refactor: Rename SlotElement -> SlottedElement
