+++ @navikt/core/react/src/slot/SlottedElement.tsx
